Refactor organizing egobox-ego with sub-modules
