+++ ego/src/egor.rs
@@ -91,9 +91,9 @@
 //!
 use crate::egor_config::*;
 use crate::errors::Result;
-use crate::mixint::*;
-use crate::solver::egor_solver::{to_xtypes, EgorSolver};
+use crate::gpmix::mixint::*;
 use crate::types::*;
+use crate::{to_xtypes, EgorSolver};
 
 use egobox_moe::GpMixtureParams;
 use log::info;
@@ -237,7 +237,7 @@ mod tests {
     use serial_test::serial;
     use std::time::Instant;
 
-    use crate::egor_solver::{DOE_FILE, DOE_INITIAL_FILE};
+    use crate::{gpmix::spec::*, DOE_FILE, DOE_INITIAL_FILE};
 
     #[cfg(not(feature = "blas"))]
     use linfa_linalg::norm::*;
