+++ ego/src/gpmix/mod.rs
@@ -1,0 +1,4 @@
+//! Mixture of Gaussian process models used by the Egor solver
+
+pub mod mixint;
+pub mod spec;
