+++ ego/src/gpmix/spec.rs
@@ -1,0 +1,7 @@
+//! GP mixture model specification used by the Egor solver
+
+/// Flags for regression model selection
+pub use egobox_moe::RegressionSpec;
+
+/// Flags for correlation model selection
+pub use egobox_moe::CorrelationSpec;
