+++ ego/src/optimizers/lhs_optimizer.rs
