+++ ego/src/solver/mod.rs
@@ -1,7 +1,7 @@
-pub mod egor_service;
-pub mod egor_solver;
-pub mod egor_state;
-pub mod solver_impl;
+mod egor_service;
+mod egor_solver;
+mod egor_state;
+mod solver_impl;
 
 pub use egor_service::*;
 pub use egor_solver::*;
